+++ Back_End/controllers/userController.js
@@ -8,7 +8,7 @@ const userController = {
     addUser: async (req, res) => {
         try {
             const { email, password, address, phoneNumber, role } = req.body;
-            const avatar = req.file ? req.file.path : null; 
+            const avatar = req.file ? req.file.path : null;
             const newUser = new User({
                 email,
                 password,
@@ -20,13 +20,13 @@ const userController = {
             const saveUser = await newUser.save();
             res.status(200).json({ EC: 0, MS: "Add user success!", saveUser });
         } catch (err) {
-            res.status(500).json({ EC: 1, MS: "Add user error!", err }); 
+            res.status(500).json({ EC: 1, MS: "Add user error!", err });
         }
     },
-    //GET ALL User
+    // GET ALL USERS
     getAllUsers: async (req, res) => {
         try {
-            const usersOld = await User.find();
+            const usersOld = await User.find({ isDeleted: false }); // Chỉ lấy người dùng không bị xóa mềm
             if (!usersOld || usersOld.length === 0) {
                 return res.status(404).json({ EC: 2, MS: "Users not found!" });
             }
@@ -45,10 +45,11 @@ const userController = {
             res.status(500).json({ EC: 1, MS: "Get all users error!", err });
         }
     },
-    //GET AN User
+    
+    // GET A USER
     getAnUser: async (req, res) => {
         try {
-            const user = await User.findById(req.params.id);
+            const user = await User.findOne({ _id: req.params.id, isDeleted: false }); // Chỉ lấy người dùng không bị xóa mềm
             if (!user) {
                 return res.status(404).json({ EC: 1, MS: "User not found" });
             }
@@ -76,7 +77,7 @@ const userController = {
                 if (user.avatar) {
                     const avatarPath = path.join(__dirname, "..", user.avatar);
                     if (fs.existsSync(avatarPath)) {
-                        fs.unlinkSync(avatarPath); 
+                        fs.unlinkSync(avatarPath);
                     }
                 }
                 user.avatar = req.file.path;
@@ -94,21 +95,31 @@ const userController = {
     //DELETE USER
     deleteUser: async (req, res) => {
         try {
-            const user = await User.findById(req.params.id);
+            const userId = req.params.id;
+            const user = await User.findById(userId);
             if (!user) {
                 return res.status(404).json({ EC: 1, MS: "User not found" });
             }
-            if (user.avatar) {
-                let avatarPath = path.join(__dirname, "..", user.avatar);
-                avatarPath = path.normalize(avatarPath);
-                if (fs.existsSync(avatarPath)) {
-                    fs.unlinkSync(avatarPath); 
-                } else {
-                    console.log(`File ${avatarPath} not found`);
+            const orders = await Order.find({ userId });
+            if (orders.length > 0) {
+                // Nếu có đơn hàng liên quan, thực hiện xóa mềm
+                user.isDeleted = true;
+                await user.save();
+                res.status(200).json({ EC: 0, MS: "User marked as deleted successfully!" });
+            } else {
+                // Nếu không có đơn hàng liên quan, thực hiện xóa cứng
+                if (user.avatar) {
+                    let avatarPath = path.join(__dirname, "..", user.avatar);
+                    avatarPath = path.normalize(avatarPath);
+                    if (fs.existsSync(avatarPath)) {
+                        fs.unlinkSync(avatarPath);
+                    } else {
+                        console.log(`File ${avatarPath} not found`);
+                    }
                 }
+                await user.remove();
+                res.status(200).json({ EC: 0, MS: "User deleted successfully!" });
             }
-            await user.remove(); 
-            res.status(200).json({ EC: 0, MS: "Delete user success!" });
         } catch (err) {
             console.error("Delete user error:", err);
             res.status(500).json({ EC: 1, MS: "Delete user error!", err });
@@ -146,11 +157,11 @@ const userController = {
             res.status(500).json({ EC: 2, MS: "Login error", err });
         }
     },
-    
+
     postSignUp: async (req, res) => {
         try {
             const { email, password, address, phoneNumber, role } = req.body;
-            const avatar = req.file ? req.file.path : null; 
+            const avatar = req.file ? req.file.path : null;
             const existingUser = await User.findOne({ email });
             if (existingUser) {
                 return res.status(400).json({ EC: 1, MS: "Email already exists." });
